[#284] feat: DocumentStatusService로 이주
